Modificari paths login

+++ AplicatiiWeb/NotesApp/views.py
@@ -1,3 +1,5 @@
+from django.contrib import messages
+from django.contrib.auth.decorators import user_passes_test
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.shortcuts import render, get_object_or_404, redirect
 from django.urls import reverse
@@ -10,12 +12,13 @@ from NotesApp.models import Notes
 # Create your views here.
 
 class NotesView(LoginRequiredMixin, ListView):
+    login_url = "../login"
     model = Notes
     template_name = 'NotesApp/notes_index.html'
     context_object_name = 'lista_notes'
 
 class CreateNotesView(LoginRequiredMixin, CreateView):
-
+    login_url = "../login"
     model = Notes
     # fields = ['name', 'text']
     form_class = NotesForm
@@ -31,6 +34,7 @@ class CreateNotesView(LoginRequiredMixin, CreateView):
 
 class UpdateNotesView(LoginRequiredMixin, UpdateView):
     model = Notes
+    login_url = "../login"
     # fields = ['name', 'text']
     form_class = NotesForm
     template_name = 'NotesApp/notes_create.html'
@@ -49,10 +53,12 @@ def delete(request, pk):
 
     if request.method == "POST":
         if 'confirm_delete' in request.POST:
-            note.delete()
-            return redirect('NotesApp:lista_notes')
-        else:
-            return redirect('NotesApp:lista_notes')
+            if request.user.is_superuser:
+                note.delete()
+                return redirect('NotesApp:lista_notes')
+            else:
+                messages.info(request, "You dont have access to delete it.")
+
 
     return render(request, 'NotesApp/confirm_delete.html', {'note': note})
